+++ IndiaEvents.Models/Models/RequestSheets/EventRequestDeviationsData.cs
@@ -10,10 +10,10 @@ namespace IndiaEvents.Models.Models.RequestSheets
     public class EventRequestDeviationsData
     {
 
-        public int HonorariumAmountExcludingTax { get; set; }
+        public double HonorariumAmountExcludingTax { get; set; }
         //public int AccommodationAmountExcludingTax { get; set; }
-        public int TravelorAccomodationAmountExcludingTax { get; set; }
-        public int OtherExpenseAmountExcludingTax { get; set; }
+        public double TravelorAccomodationAmountExcludingTax { get; set; }
+        public double OtherExpenseAmountExcludingTax { get; set; }
         //public int FoodAndBeveragesAmountExcludingTax { get; set; }
         //public int AgregateSpendAmountExcludingTax { get; set; }
         public string? DeviationFile { get; set; }
